fix(opinion_page): address Alberto's follow-up review on SCOTUS docket rendering
- Move get_docket_entry_source into Docket.get_entry_source()
- Build scotus_metadata in core_docket_data() so it's available on
  every docket tab
- Render the SCOTUS metadata section with a for loop over the
  pre-built item list, instead of manual per-field blocks
- Format date_discretionary_court_decision with Django's date_format()
  instead of strftime()
- Prioritize questions_presented_file over questions_presented_url
- Move build_scotus_metadata into docket_entry_sources.py and rename
  it to docket_sources_utils.py
- Remove hide_docket_alerts
- Link to the SCOTUS coverage page (pending #7368)
