+++ cl/opinion_page/templates/includes/de_filter.html
@@ -111,7 +111,11 @@
 
 <p style="margin-top: 10px; margin-bottom: 0;">
   {% if is_scotus %}
-    <i class="fa fa-info-circle"></i> Items on this page are sourced from the Supreme Court of the United States and may not be up to date.
+    {% comment %}
+    Swap this hardcoded path for {% url 'coverage_scotus' %} once the
+    SCOTUS coverage page is ready.
+    {% endcomment %}
+    <i class="fa fa-info-circle"></i> Items on this page are sourced from the Supreme Court of the United States and may not be up to date. <a href="/help/coverage/scotus/">Learn more about our coverage</a>.
   {% else %}
     <i class="fa fa-info-circle"></i> Items on this page are sourced from <a href="https://www.pacer.gov">PACER</a> and <a href="https://free.law/recap/">RECAP</a> and may not be up to date. <a href="{% url 'coverage_recap' %}">Learn more about our coverage</a>.
   {% endif %}
